+++ CEffect_DamageNumber.h
@@ -6,10 +6,10 @@ class CEffect_DamageNumber :
     public CEffect
 {
     int cnt = 0;
-    int num;
+    double num;
     int type = 0;
 public:
-    CEffect_DamageNumber(CVector position, int num, int type);
+    CEffect_DamageNumber(CVector position, double num, int type);
 
     void Update();
     void Render()const;
