+++ CStatus.h
@@ -3,9 +3,9 @@ class CStatus
 {
 public:
 	int Level;
-	int ATK;
-	int DEF;
-	int HP;
+	double ATK;
+	double DEF;
+	double HP;
 
 	CStatus(int Level);
 	CStatus(int Level, int ATK, int DEF, int HP);
